[Perf] Address review: merge cuda_ipc into gpu_ipc_memory; use torch.accelerator
