+++ vllm/v1/engine/core.py
@@ -82,7 +82,7 @@ from vllm.v1.kv_cache_interface import KVCacheConfig, get_kv_cache_spec_kind
 from vllm.v1.metrics.stats import SchedulerIterationDetails, SchedulerStats
 from vllm.v1.outputs import ModelRunnerOutput
 from vllm.v1.request import Request, RequestStatus
-from vllm.v1.serial_utils import MsgpackDecoder, MsgpackEncoder
+from vllm.v1.serial_utils import MsgpackDecoder, MsgpackEncoder, OOBTensorProvider
 from vllm.v1.structured_output import StructuredOutputManager
 from vllm.v1.utils import compute_iteration_details
 from vllm.version import __version__ as VLLM_VERSION
@@ -1025,7 +1025,7 @@ class EngineCoreProc(EngineCore):
         # Out-of-band provider for multimodal tensor IPC. torch_shm drains a
         # queue; cuda_ipc returns a deferred proxy that each worker materializes.
         self.tensor_ipc_receiver: TensorIpcReceiver | None = None
-        self._oob_provider = None
+        self._oob_provider: OOBTensorProvider | None = None
         if tensor_queue is not None:
             self.tensor_ipc_receiver = TensorIpcReceiver(tensor_queue)
             self._oob_provider = self.tensor_ipc_receiver
@@ -1033,7 +1033,7 @@ class EngineCoreProc(EngineCore):
         else:
             mm_config = getattr(vllm_config.model_config, "multimodal_config", None)
             if mm_config is not None and mm_config.mm_tensor_ipc == "cuda_ipc":
-                from vllm.multimodal.cuda_ipc import cuda_ipc_provider
+                from vllm.multimodal.gpu_ipc_memory import cuda_ipc_provider
 
                 self._oob_provider = cuda_ipc_provider
                 logger.info("Using CUDA-IPC pool for multimodal tensor sharing")
